+++ Live2DAction/Assets/_Project/Game/Camera/CameraViewMode.cs
@@ -1,8 +1,0 @@
-namespace Live2DAction.CameraSystem
-{
-    public enum CameraViewMode
-    {
-        ThirdPerson,
-        FirstPerson
-    }
-}